scripts/requirements-vc-enrich.txt
# pip install -r scripts/requirements-vc-enrich.txt
google-generativeai>=0.8.0
requests>=2.31.0
